+++ self-host/src/components/layout-components/components/channel-layout.tsx
